qcacld-2.0: Pass tx_pkt_fail_cnt_threshold to FW-1
Add data structures to save tx_pkt_fail_cnt_threshold and changes
to pass the same to FW.

CRs-Fixed: 1041575

+++ CORE/MAC/inc/sirApi.h
@@ -7494,7 +7494,7 @@ struct sme_update_access_policy_vendor_ie {
  */
 struct sme_tx_fail_cnt_threshold {
 	uint8_t session_id;
-	uint16_t tx_fail_cnt_threshold;
+	uint32_t tx_fail_cnt_threshold;
 };
 
 /**
